backend/requirements.txt
fastapi==0.115.2
uvicorn==0.30.1
pydantic==2.9.2
azure-cosmos==4.6.0
